fix(web): render GFM tables in the document editor
The Tiptap document editor's schema had no table nodes, so tiptap-markdown
parsed GFM pipe tables into <table> HTML that ProseMirror silently dropped,
flattening cell text into run-together paragraphs. Any edit then autosaved
the corrupted doc back to disk, destroying the tables in the source file.

Register the @tiptap/extension-table nodes and extract the extension list
into a shared module with headless parse and round-trip tests.
